commit de todos los objetos terminados y el program con comandos para test

+++ src/Library/Dish.cs
@@ -12,10 +12,10 @@ namespace Ucu.Poo.Restaurant
     public class Dish
     {
         public string Name { get; set; }
-        public decimal Price { get; set; }
+        public double Price { get; set; }
         public bool IsVegetarian { get; set; }
     
-        public Dish(string name, decimal price, bool isVegetarian)
+        public Dish(string name, double price, bool isVegetarian)
         {
             Name = name;
             Price = price;
